pipeline: Save the correct trainer object.
Make sure that the correct Trainer object is save after each experiment.
For the attacks sb.trainer and <attack_class>.trainer is the same object
so there were no issues until now.

However, for the defenses there are two different trainers that we need
to take care. The first is the trainer that is loaded from the disk and
contains the poisoned model and the other is the randomly initialized
Trainer in the SystematicBackdoor class. In the defense class the
correct trainer is used (the one with the poisoned model) but when the
defense finishes its execution the sb.trainer is an unused object that
is randomly initialized and in main.py this will be saved to disk. For
this reason after executing the defense we need to overwrite the trainer
in SystematicBackdoor class with the correct Trainer object, ensuring
that the correct object is saved. Additionally, we need to update this
trainer's id with the defense's id to avoid overwriting the folder with
the original poisoned model. For this reason use an id parameter in the
Trainer class.

Apply also some formatting changes.

Signed-off-by: Stefanos Koffas <koffas.stefanos@gmail.com>